requirement.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
idna==3.11
pillow==12.1.0
python-decouple==3.8
razorpay==2.0.0
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
